Fixes #3466. Refs #3410. This fixes preview for XVideo chromkeyed video in the program guide video preview caused by the fix for #3410, and fixes a longstanding bug that caused the video to 'flash' with other video renderers in the preview.
The fix for the DrawUnusedRects() being drawn in the wrong place in #3410 disabled the drawing of the chromakey color causing chromakeyed video to be hidden from view. This patch instead makes the guidegrid call DrawUnusedRects() when it is thread safe to do so.

It turned out that DrawUnusedRects() was also causing flashing whenever there was a prebuffering pause during playback, so this fixes that problem as well.

git-svn-id: http://svn.mythtv.org/trunk@13617 4b91aca1-2a1f-0410-a73a-a78f2df169e8