Fixes #3401. Applies slightly modified version of patch.
This just checks if we have the xrandr extension before using it. This fixes a segfault for people recompile their own X11 servers to exclude XRandR support after compiling MythTV, or when they have X11 installed on an OSX machine.

I added the dummy variables argued over in the ticket comments. While I don't think they are required by the API, an implementation of XRandR in wide release appears to require them in practice, whether this is a new bug in the implementation or a new bug in the API isn't really important; the code that works with this XRandR release is still correct according API and only one line longer.

git-svn-id: http://svn.mythtv.org/trunk@13766 4b91aca1-2a1f-0410-a73a-a78f2df169e8