Adds a date filter to the events screen to show only events from a given date.
Add two new settings to allow the date and time formats used on the events
screen to be changed.

Try flipping the meaning of up/down and left/right key presses that are handled
by the selectors to see what sort of reaction it gets. This means left/right
consistently move focus between widgets and up/down changes selection in the
selectors. That's contrary to the way selectors work in other parts of myth
but would seem to be a more consistent way of doing things.

NOTE: changes the protocol version in mythzmserver so both the server and all 
clients need to be updated at the same time.



git-svn-id: http://svn.mythtv.org/trunk@13843 4b91aca1-2a1f-0410-a73a-a78f2df169e8