Extends the audio, video and subtitle information we recognise about a program by replacing the current hdtv, stereo and closecaptioned flags with a videoproperties, audioproperties and subtitletype vars. The changes should be backwards compatible with existing program/recordprogram table values. For now the stereo, hdtv and closecaptioned columns haven't been renamed.
The following values are now accepted and can be extended to support more e.g. Radio programs could be indicated by  VID_NONE:
{{{
enum AudioProps {
    AUD_UNKNOWN       = 0,
    AUD_STEREO        = 1,
    AUD_MONO          = 2,
    AUD_SURROUND      = 3,
    AUD_DOLBY         = 4
};

enum VideoProps {
    VID_UNKNOWN       = 0,
    VID_HDTV          = 1,
    VID_WIDESCREEN    = 2
};

enum SubtitleTypes {
    SUB_UNKNOWN       = 0,
    SUB_HARDHEAR      = 1,
    SUB_NORMAL        = 2,
    SUB_ONSCREEN      = 3
};
}}}

These values are in line with the information made available through Xmltv and EIT.

The xmltv parser has been changed to use the information if it is available. DataDirect and EIT are currently unaffected.

Although no themes yet support it, the icons for the watch recordings screen are already in SVN. I'll update the themes as I get the time. 



git-svn-id: http://svn.mythtv.org/trunk@13952 4b91aca1-2a1f-0410-a73a-a78f2df169e8