Add initial support to the UK EIT fixup for the "Starring <actor 1> and <actor 2>, <year>" information that has begun appearing in the EIT data.
For the first time this gives UK EIT users
some actor information that we've never had
before.

This appears to be correct about 97% of the time.

Additionally, the airdates that are found are
propagated to the airdate field as well as the
origairdate field. This allows mythweb to
correctly show the original showing year
for movies.


git-svn-id: http://svn.mythtv.org/trunk@13970 4b91aca1-2a1f-0410-a73a-a78f2df169e8