Refs #3727. Adds SA 4250HDC STB to recognized list but tells user that it is not supported.
Unfortunately none of the existing firewire channel changing code works with the SA 4250 HDC, but by adding it to the recognized list we can at least tell the user that their STB is not supported. You can still use this STB using an IR blaster, but then you have the reliability issues that IR blaster can have.

git-svn-id: http://svn.mythtv.org/trunk@14110 4b91aca1-2a1f-0410-a73a-a78f2df169e8