The offset argument to mythfilldatabase --file doesn't actually do anything, so for the sake of clarity I'm removing it.
Any scripts which are using the --file argument to import guide data manually will need to be updated. This might be a good time to check if you still need to use it. Most grabbers are now supported directly by mythfilldatabase and the --file method should only be used in a few unusual cases.



git-svn-id: http://svn.mythtv.org/trunk@14123 4b91aca1-2a1f-0410-a73a-a78f2df169e8