Add rename feature for MythGallery files and directories.
Based on the original patch by Guido Grazioli.

Changes by me to ensure the metadata in the database is updated properly and
renames any associated directory and file thumbnails so we don't leave
stale thumb images behind. Tell the user if the rename failed and I also 
tweaked the default theme file a little to make room for the new menu item.

Closes #3713.

git-svn-id: http://svn.mythtv.org/trunk@14208 4b91aca1-2a1f-0410-a73a-a78f2df169e8