Fixes a channel scanning for DVB-S cards that support probing tuning information from the hardware.
This is a quick fix which shouldn't break anything and will fix a serious problem for DVB-S scans with many cards. We were returning the raw frequency used by the tuning hardware on the card, and not the actual carrier frequency in ProbeTuningParams in the presence of an LNB. We need an inverse of DiSEqCDevLNB::GetIntermediateFrequency() to properly support ProbeTuningParams() for these cards. Since such a function is a complex beast, I've just disabled the optional ProbeTuningParams() for these cards for now.

git-svn-id: http://svn.mythtv.org/trunk@14359 4b91aca1-2a1f-0410-a73a-a78f2df169e8