Adds auto discovery of master backend servers to mythfrontend.
Be careful when running more than one server with different versions.
If you choose a server with an older version of Myth on it, it will
upgraded the database without warning.

I ran out of time so I just commented out the lirc logic.  
Support for remotes in the master selection and pin screens, will be 
added at a later date.

See ticket # 3293 for more details http://svn.mythtv.org/trac/ticket/3293

Closes #3293



git-svn-id: http://svn.mythtv.org/trunk@14492 4b91aca1-2a1f-0410-a73a-a78f2df169e8