Allow mythbackend's --printexpire command line option to take an optional hostname as an argument.  If a hostname is given, only expirable recordings that were made on that particular backend are listed.
I don't see a whole lot of reason to show this information in any other place
because the hostname by itself isn't all that valuable in large Myth
installations that utilize multiple storage directories or share the same
directories between multiple hosts.  It only took a few minutes to implement
this much though, so I decided to go ahead and do it.

Closes #3802.

This does not address the comment on the ticket about the Auto-Expirer trying
to delete programs on non-connected backends.  The current code does check to
see if a backend is connected before it determines a file is expireable and
tries to expire it.  If this issue is reproduceable, please provide more
debugging information in a separate ticket as it is unrelated to this one.



git-svn-id: http://svn.mythtv.org/trunk@14547 4b91aca1-2a1f-0410-a73a-a78f2df169e8