Refs #4009. Places videoout helper enums and typedefs that are used throughout MythTV in their own file videoouttypes.h.
This changeset looks bigger than it is, because this uses enums throughtout where the enums had been cast to int before.

When looking at the #4009, aspect ratio issue the other day I noticed that the translation strings for these are created in multiple places because it is not safe to include videooutbase.h everywhere. And we are also passing these around as integers and hard defining the MAX value of the enums throughout the code. The reduntant translations are error prone and the use of hard coded MAX values is dangerous when new items are added or removed from an enum.

The new videoouttypes.h has been added to the translate.pro and some strings that should have been translated before but were in files that aren't in translate.pro will now be availabe for translation.



git-svn-id: http://svn.mythtv.org/trunk@14592 4b91aca1-2a1f-0410-a73a-a78f2df169e8