Closes #3973
This patch was contributed by Doug Young (and modified by me, any bugs 
or other brokeness are mine).

 * Added ability for a unique PIN to protect parental level 2 and 3.

 * Disallow default parental level view without a PIN (if required).

'''Caution:''' A functionality change in this patch:

Before this patch a PIN (if set) was required for any level change.
The new PIN levels currently allow for access to levels 2 and 3 without
a PIN (by default). If you want the old behavior you need to set
"Parental Level 2 PIN".

This can be changed if it turns out to be a bad idea.

PIN checking:

For a level all PINs for that level and up are allowed. If no PINs are
found in the first direction, the first valid PIN at a lower level is used.
No PINs = all access.



git-svn-id: http://svn.mythtv.org/trunk@14705 4b91aca1-2a1f-0410-a73a-a78f2df169e8