Add a new 'DVD Drive Write Speed" setting to MythArchive.
Allows the maximum write speed to be specified which is usefull if you have
some cheap media that you need to slow down the write speed to get good 
results from.

NOTE: It's not guaranteed that growisofs will actually use the speed you
specify it depends on the capabilities of the drive and the media being used.

I got some strange results when I tried it. Setting a speed of 2x actually
appears to have burned the DVD at 4.1x and 8x burned at 16.1x.


git-svn-id: http://svn.mythtv.org/trunk@14761 4b91aca1-2a1f-0410-a73a-a78f2df169e8