Refs #4115. Fixes segfault in "All Tuners are Busy" dialog handling and adds an "Exit" option.
I noticed two problems with the handling of this dialog when debugging the network control.
 * If you hit escape in this dialog it would select the first recording in the list for playback, rather than returning to the main menu.
 * If you send the "return" key to the frontend application when the frontend is not focused it segfaults.

The first problem occurs because the "reject" action of the showButtonPopup() function returns '0' the index of the first recording until this changeset added an exit option. The second problem occurs because there was no range checking on the return from showButtonPopup(). We really need to move to the new less buggy MythUI dialogs in the long term, but a little bit of range checking and adding an "Exit" option to this dialog are good ideas anyway.



git-svn-id: http://svn.mythtv.org/trunk@14777 4b91aca1-2a1f-0410-a73a-a78f2df169e8