Fixes #4096. Channel list was empty after a certain series of actions.
The problem was not in the channel editor but in the ListBoxSetting itself which deleted the reference to the current MythListBox widget after configWidget was called the second time. Also clearSelections() did not properly clear the list if there already was a widget. This fixes both problems.

This change requires recompiling plug-ins due to the updated binary version for change in the ListBoxSetting signature.

git-svn-id: http://svn.mythtv.org/trunk@14791 4b91aca1-2a1f-0410-a73a-a78f2df169e8