Refs #4115. Fixes some more Network Control segfaults.
This fixes segfaults like the second one described in commit [14858]. This happens anytime a signal arrives for slot in a dialog that has already been deleted. This can happen without Network Control, but is very rare in that case. With network control it is easy to send two commands back to back or send a key while the UI is not focused.

This only addresses this type of segfault for all uses of the MythDialog class. I haven't had time to test the other dialogs.

git-svn-id: http://svn.mythtv.org/trunk@14862 4b91aca1-2a1f-0410-a73a-a78f2df169e8