Refs #4115. Refs #4132. Changes explicit deletes of three dialogs to deleteLater() calls.
Like [14862] which addressed this problem for MythDialog, this addresses the delete problem for three children of MythDialog, MythProgressDialog, MythBusyDialog, and DialogBox. As explained in the comments on [14858] and [14862] deleting these dialogs explicitly can allow signals to arrive for the dialog after it has been deleted causing segfaults. This is hard to do when using a keyboard and mouse to provide input, but much easier to trigger when using network control to send key commands to the UI.

This only fixes 3 dialogs and the change appears much larger than it in fact is. I'm not doing all the dialogs at once so that I can roll back a just a few at a time if there are any problems. If you have an outside plugin that fails to compile after this changeset, you probably just need to change some 'delete dialogname;' expressions to 'dialogname->deleteLater();'.

This does change the myth binary revision so plugins will have to be recompiled.

git-svn-id: http://svn.mythtv.org/trunk@14874 4b91aca1-2a1f-0410-a73a-a78f2df169e8