Fixes #4162. Adjusts images in mythgallary to the correct aspect ratio.
This adds a function to util.h MythGetPixelAspectRatio() which returns the aspect ratio of the default screen's pixel aspect ratio and uses this to adjust the display of the image. This will not work properly if there are multiple screens with different pixel aspect ratios, and it only works on X11 platforms (but this would not be too difficult to port).

To support the util.h function three functions have been added to util-x11.h, MythXGetDisplaySize() which determines a screen's size in pixels, MythXGetDisplayDimensions(), which returns a sreen's size in millimeters, and MythXGetPixelAspectRatio() which determines a screen's pixel aspect ratio using those two functions. These functions accept a Display* and a screen number, but when called without them use MythXOpenDisplay() to open the correct display and use DefaultScreen() to select the screen number.DisplayResX and videoout_xv.cpp now use these functions to query the screen rather than duplicating the code.

git-svn-id: http://svn.mythtv.org/trunk@14950 4b91aca1-2a1f-0410-a73a-a78f2df169e8