Fixes #3478. Quiet logs about invalid dates when parsing programinfo.
This patch from mtdean just checks for empty and '0000-00-00' dates in ProgramInfo::FromStringList() before parsing them, and set's them to a null date in this case, rather than having QDateTime do it for us and print a warning. The dates can be null on a normally functioning MythTV system so we shouldn't be printing these warnings constantly.

git-svn-id: http://svn.mythtv.org/trunk@14953 4b91aca1-2a1f-0410-a73a-a78f2df169e8