Refs #3872. Do a firewire bus reset when required in the LinuxFirewireDevice.
This is the default behaviour with the DarwinFirewireDevice, but due to concerns about buggy Linux drivers this requires changing a line of code and recompiling in Linux to enable. But since there are many bug reports which are really do to this not being enabled, this changeset enables the reset and adds an option to mythtv-setup+general|miscellaneous to disable it.

The problem drivers were for an external firewire harddrive, they have been fixed in the Linux kernel for over a year. Resets can also cause a small glitch in recordings currently in progress if you have multiple Motorolla STBs connected to the same firewire bus, but a this is preferable to most people over zero-byte recordings.

git-svn-id: http://svn.mythtv.org/trunk@14956 4b91aca1-2a1f-0410-a73a-a78f2df169e8